requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
altair==4.2.2
appdirs==1.4.4
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.1.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
clikit==0.6.2
crashtest==0.3.1
entrypoints==0.4
flatbuffers==25.2.10
frozenlist==1.5.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
grpcio==1.70.0
h5py==3.13.0
holidays==0.35
httpstan==4.10.1
idna==3.10
importlib-metadata==6.11.0
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==2.15.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
ml-dtypes==0.2.0
multidict==6.1.0
numpy==1.26.1
oauthlib==3.2.2
opt_einsum==3.4.0
packaging==23.2
pandas==2.1.1
pastel==0.2.1
Pillow==9.5.0
propcache==0.3.0
protobuf==4.25.6
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydeck==0.9.1
Pygments==2.19.1
pylev==1.4.0
pysimdjson==5.0.2
pystan==3.7.0
python-dateutil==2.9.0.post0
pytz==2025.1
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.23.1
rsa==4.9
scikit-learn==1.3.2
scipy==1.15.2
six==1.17.0
smmap==5.0.2
streamlit==1.27.0
tabulate==0.9.0
tenacity==8.5.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
threadpoolctl==3.5.0
toml==0.10.2
toolz==1.0.0
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3
urllib3==2.3.0
validators==0.34.0
watchdog==6.0.0
webargs==8.6.0
Werkzeug==3.1.3
wrapt==1.14.1
yarl==1.18.3
zipp==3.21.0
